requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.17.1
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.100.0
anyio==4.9.0
APScheduler==3.10.4
asttokens==3.0.0
astunparse==1.6.3
attrs==25.1.0
av==16.0.1
backoff==1.11.1
base58==2.1.1
bcrypt==5.0.0
beanie==1.28.0
beautifulsoup4==4.12.3
bidict==0.23.1
billiard==4.2.4
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
celery==5.4.0
certifi==2025.1.31
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.21.1
colorama==0.4.6
coloredlogs==15.0.1
comfyui-embedded-docs==0.3.1
comfyui_frontend_package==1.28.8
comfyui_workflow_templates==0.2.11
comm==0.2.2
confection==0.1.5
ConfigArgParse==1.7.5
construct==2.10.68
construct-typing==0.5.6
contourpy==1.3.1
coverage==7.13.5
cryptography==46.0.7
cycler==0.12.1
cymem==2.0.11
dash==4.1.0
debugpy==1.8.12
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.18.0
docutils==0.21.2
ecdsa==0.19.2
einops==0.8.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
eventlet==0.40.2
executing==2.2.0
fastapi==0.115.6
fastjsonschema==2.21.2
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.1.24
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.5.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.10.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.1
groq==0.31.1
grpcio==1.76.0rc1
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
h5py==3.12.1
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.33.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
importlib_metadata==9.0.0
iniconfig==2.3.0
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.11.0
joblib==1.5.1
jsonalias==0.1.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.16
keras==3.10.0
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.sdl2==0.8.0
kiwisolver==1.4.8
kombu==5.6.2
kornia==0.8.1
kornia_rs==0.1.9
langcodes==3.5.0
language_data==1.3.0
lazy-loader==0.5
lazy-model==0.2.0
libclang==18.1.1
m3u8==6.0.0
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mmh3==5.2.1
motor==3.6.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.13
mypy==1.14.1
mypy_extensions==1.1.0
namex==0.0.8
narwhals==1.45.0
nbformat==5.10.4
neo4j==5.27.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
open3d==0.19.0
openai==1.108.1
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pillow==11.1.0
platformdirs==4.3.6
plotly==6.6.0
pluggy==1.6.0
postgrest==2.30.0
preshed==3.0.10
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==6.1.1
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.13.2
pydantic-settings==2.11.0
pydantic_core==2.46.2
pydeck==0.9.1
pyfiglet==1.0.2
PyGetWindow==0.0.9
Pygments==2.19.1
pyiceberg==0.11.1
PyJWT==2.12.1
pymongo==4.9.2
PyMsgBox==1.0.9
pyparsing==3.2.1
pyperclip==1.9.0
pypiwin32==223
pyreadline3==3.5.4
PyRect==0.2.0
pyroaring==1.1.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.2
python-jose==3.3.0
python-multipart==0.0.20
python-socketio==5.13.0
python-telegram-bot==21.4
pytweening==1.2.0
pytz==2025.1
pywhatkit==5.4
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
qrcode==8.0
realtime==2.30.0
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
reportlab==4.3.1
requests==2.32.5
requests-oauthlib==2.0.0
retrying==1.4.2
rich==13.9.4
rpds-py==0.26.0
rsa==4.9
ruff==0.8.6
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.7.0
scipy==1.16.0
segno==1.6.1
selenium==4.28.1
sentencepiece==0.2.1
setuptools==75.8.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
solana==0.35.0
solders==0.21.0
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spandrel==0.4.1
spotify==0.10.2
SQLAlchemy==2.0.40
srsly==2.5.1
stack-data==0.6.3
starlette==0.41.3
storage3==2.30.0
streamlit==1.46.1
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.30.0
supabase-auth==2.30.0
supabase-functions==2.30.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.19.0
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2026.3.3
tokenizers==0.21.2
toml==0.10.2
torch==2.8.0
torchaudio==2.8.0
torchsde==0.2.6
torchvision==0.23.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trampoline==0.1.2
transformers==4.53.2
trimesh==4.11.5
trio==0.28.0
trio-websocket==0.11.1
typer==0.16.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.3.0
uv==0.7.19
uvicorn==0.34.0
vine==5.1.0
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.13
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.15
wikipedia==1.4.0
wrapt==1.17.2
wsproto==1.2.0
xlrd==2.0.2
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
